refactor: extract getTooltips function into its own module for improved organization and maintainability
